Remove binder from transpileDeclaration implementation.
