+++ src/compiler/_namespaces/ts.ts
@@ -57,7 +57,6 @@ export * from "../transformers/module/system";
 export * from "../transformers/module/esnextAnd2015";
 export * from "../transformers/module/node";
 export * from "../transformers/declarations/diagnostics";
-export * from "../transformers/declarations/emitBinder";
 export * from "../transformers/declarations/emitResolver";
 export * from "../transformers/declarations/transpileDeclaration";
 export * from "../transformers/declarations/localInferenceResolver";
