+++ src/compiler/transformers/declarations/localInferenceResolver.ts
@@ -19,7 +19,6 @@ import {
     GetAccessorDeclaration,
     getCommentRange,
     getEmitScriptTarget,
-    getMemberKeyFromElement,
     getNameOfDeclaration,
     getTextOfNode,
     HasInferredType,
@@ -498,7 +497,7 @@ export function createLocalInferenceResolver({
                 }
             }
 
-            const nameKey = getMemberKeyFromElement(prop);
+            const nameKey = prop.symbol.escapedName;
             const name = normalizePropertyName(prop.symbol, isMethodDeclaration(prop)) ??
                 deepClone(visitNode(prop.name, visitDeclarationSubtree, isPropertyName)!);
 
@@ -530,7 +529,6 @@ export function createLocalInferenceResolver({
                 });
 
                 if (nameKey) {
-                    Debug.assertSymbolValid(prop.symbol);
                     const exitingIndex = members.get(prop.symbol);
                     if (exitingIndex !== undefined && !isMethodOrAccessor(prop)) {
                         properties[exitingIndex] = newProp;
@@ -664,7 +662,6 @@ export function createLocalInferenceResolver({
 
     function normalizePropertyName(symbol: Symbol, isMethod: boolean) {
         let nameText;
-        Debug.assertSymbolValid(symbol);
         Debug.assert(symbol.declarations !== undefined, "Symbol has no declarations");
         let stringNamed = !!length(symbol.declarations);
         let singleQuote = stringNamed;
@@ -790,7 +787,6 @@ export function createLocalInferenceResolver({
             return localInference(node.expression, NarrowBehavior.KeepLiterals);
         }
         else if (isVariableDeclaration(node)) {
-            Debug.assertSymbolValid(node.symbol);
             const firstDeclaration = node.symbol.valueDeclaration;
             // Use first declaration of variable for the type
             if (node !== firstDeclaration && firstDeclaration && isVariableDeclaration(firstDeclaration)) {
